Move descriptors out of detail
