+++ core/include/detray/geometry/surface_descriptor.hpp
