+++ core/include/detray/geometry/volume_descriptor.hpp
