+++ core/include/detray/navigation/accelerators/surface_grid.hpp
@@ -8,7 +8,7 @@
 #pragma once
 
 // Project include(s)
-#include "detray/geometry/detail/surface_descriptor.hpp"
+#include "detray/geometry/surface_descriptor.hpp"
 #include "detray/utils/grid/detail/axis_helpers.hpp"
 #include "detray/utils/grid/detail/grid_bins.hpp"
 #include "detray/utils/grid/grid.hpp"
