+++ tests/integration_tests/cpu/propagator/covariance_transport.cpp
@@ -7,10 +7,10 @@
 
 // Project include(s).
 #include "detray/definitions/units.hpp"
-#include "detray/geometry/detail/surface_descriptor.hpp"
 #include "detray/geometry/mask.hpp"
 #include "detray/geometry/shapes.hpp"
 #include "detray/geometry/shapes/unbounded.hpp"
+#include "detray/geometry/surface_descriptor.hpp"
 #include "detray/navigation/intersection/helix_intersector.hpp"
 #include "detray/propagator/detail/jacobian_engine.hpp"
 #include "detray/test/framework/types.hpp"
