+++ tutorials/include/detray/tutorial/detector_metadata.hpp
@@ -12,8 +12,8 @@
 #include "detray/core/detail/single_store.hpp"
 #include "detray/definitions/containers.hpp"
 #include "detray/definitions/indexing.hpp"
-#include "detray/geometry/detail/surface_descriptor.hpp"
 #include "detray/geometry/mask.hpp"
+#include "detray/geometry/surface_descriptor.hpp"
 #include "detray/io/backend/detail/type_info.hpp"  // mask_info
 #include "detray/materials/material_slab.hpp"
 #include "detray/navigation/accelerators/brute_force_finder.hpp"
